Split pull request slash parser
