+++ docs/DECISIONS.md
@@ -236,6 +236,7 @@
 - Runtime/auth status labels live in `QuillCodeRuntimeStatusLabel`, separate from lifecycle labels. `RuntimeFactory`, `WorkspaceRuntimeIssueBuilder`, and the desktop sign-in controller should use those constants for mock, sign-in-needed, developer-key-needed, signed-in, ready, and sign-in-failed statuses so runtime issue detection does not depend on scattered string sentinels.
 - Agent progress status copy lives in `WorkspaceAgentStatusBuilder`. `WorkspaceModel` owns when agent progress arrives and when the top bar refreshes, while the builder owns event-kind-to-status labels such as Queued, Running, Review, Streaming, Finishing, and Failed so progress copy stays directly tested and aligned with top-bar presentation semantics.
 - Slash-command local transcript copy lives in `WorkspaceSlashCommandTranscriptPlanner`. `WorkspaceModel` still owns command side effects such as renaming, scheduling, SSH project creation, memory writes, local environment action execution, and tool dispatch, but the user-visible success/failure records for `/help`, `/status`, `/mode`, `/model`, `/rename`, `/project rename`, `/ssh`, `/follow-up`, `/workspace-check`, `/env`, invalid commands, and unknown commands are pure planner outputs with focused tests.
+- Pull request slash-command parsing lives in `SlashPullRequestCommandParser`. `SlashCommandParser` owns only the top-level `/pr` delegation, while PR subcommands, selector/body splitting, reviewer and label argument construction, and merge flags stay beside focused parser tests.
 - Memory command orchestration lives in `WorkspaceMemoryEngine`, not inline in `WorkspaceModel`. Transcript copy lives in `WorkspaceMemoryCommandTranscriptPlanner`, write/delete error copy goes through `WorkspaceMemoryErrorMessageBuilder`, and thread refresh events go through `WorkspaceMemoryContextUpdatePlanner`. The engine owns global memory write/delete outcomes, global reloads, transcript mutations, and context-update intents; `WorkspaceModel` owns only applying those intents to actor-isolated workspace state, persistence, and top-bar refreshes.
 - Per-turn agent runner configuration lives in `WorkspaceAgentRunContextBuilder`. `WorkspaceModel` selects the current project/browser/memory/MCP/Computer Use state and asks the builder for a configured `AgentRunner`; the builder owns local versus SSH Remote base tools, optional plan/browser/Computer Use/memory/MCP tool definitions, and override composition. Memory tool execution and saved-memory event detection live together in `WorkspaceMemoryRememberToolExecutor`.
 - Per-turn agent send execution lives in `WorkspaceAgentSendSession`. `WorkspaceModel` still owns composer state, runner-context selection, thread persistence, and top-bar side effects, while the session owns cancellation checks, the `AgentRunner.send` call, progress handler handoff, and the saved-memory signal returned from the completed thread.
